moved addon classes to components folder
moved addon classes to components folder
